테스트 추가

+++ sw_all/src/sw_all/Main_Memo.java
@@ -27,19 +27,19 @@ public class Main_Memo {
                   Create_m();
                   break;
                case 2:
-                  Scanner s1 = new Scanner(System.in);
-                  if (View2_m()>0) {
-                     System.out.println("특정 메모의 내용을 보시겠습니까?(yes인경우: yes 입력후 엔터치고 메모의 제목을 입력하세요, no인 경우:no를 입력하세요)");
-                      String ans = s1.nextLine();
-                      if( ans.equals("yes")) {
-                         String title = s1.nextLine();
-                         View_m(title);
-                      }
-                  }
-                  else {
-                      System.out.println("생성된 메모가 존재하지 않습니다.");
+                   Scanner s1 = new Scanner(System.in);
+                   if (View2_m()>0) {
+                      System.out.println("특정 메모의 내용을 보시겠습니까?(yes인경우: yes 입력후 엔터치고 메모의 제목을 입력하세요, no인 경우:no를 입력하세요)");
+                       String ans = s1.nextLine();
+                       if( ans.equals("yes")) {
+                          String title = s1.nextLine();
+                          View_m(title);
+                       }
                    }
-                  break;
+                   else {
+                       System.out.println("생성된 메모가 존재하지 않습니다.");
+                    }
+                   break;
                case 3:
                   Scanner s2 = new Scanner(System.in);
                   System.out.println("수정할 메모의 제목을 입력하세요");
@@ -58,42 +58,47 @@ public class Main_Memo {
             }
          }
    }
+static public boolean createMemo(String title) {
+	File file1 = new File("c://Example//File//"+ title+ ".txt");
+	return true;
+}
 
-   static void Create_m(){
-       BufferedWriter writer = null;
-       Scanner s2 = new Scanner(System.in);
-       Scanner s3 = new Scanner(System.in);
-       Scanner s4 = new Scanner(System.in);
-              try {
-            	  System.out.println("날짜를 입력하세요(ex,2018년 11월 20일 오후 2시 20분의 경우: 2018 11 20 14 20로 입력함)");
-                  String date = s3.nextLine();
-                  System.out.println("제목을 입력하세요");
-                  String title = s2.nextLine();
-                  File file1 = new File("c://Example//File//");
-                  File[] files = file1.listFiles();
-                  for (int fileIntList = 0; fileIntList < files.length; fileIntList++) {
-                	  String ss = files[fileIntList].toString();
-                	  if (title.equals(ss.substring(ss.lastIndexOf("\\") + 1, ss.length()-4))) {
-                		  System.out.println("다른 제목을 입력하세요.");
-                		  title = s4.nextLine();
-                	  }
-                  }
-                 writer = new BufferedWriter(new FileWriter("c://Example//File//"+ title + ".txt"));
-                 writer.write("날짜:" + date + "\r\n");
-                 writer.write("제목:" + title + "\r\n");
-                  System.out.println("내용을 입력하세요");
-                  String story = s.nextLine();
-                  while (!story.equals("store")){
-                     writer.write( story + "\r\n");
-                     story = s2.nextLine();
-                  }
-                  writer.close();
-              } catch (FileNotFoundException e) {
-                  e.printStackTrace();
-              } catch (IOException e) {
-                  e.printStackTrace();
-              }
-          }
+
+      static void Create_m(){
+          BufferedWriter writer = null;
+          Scanner s2 = new Scanner(System.in);
+          Scanner s3 = new Scanner(System.in);
+          Scanner s4 = new Scanner(System.in);
+                 try {
+               	  System.out.println("날짜를 입력하세요(ex,2018년 11월 20일 오후 2시 20분의 경우: 2018 11 20 14 20로 입력함)");
+                     String date = s3.nextLine();
+                     System.out.println("제목을 입력하세요");
+                     String title = s2.nextLine();
+                     File file1 = new File("c://Example//File//");
+                     File[] files = file1.listFiles();
+                     for (int fileIntList = 0; fileIntList < files.length; fileIntList++) {
+                   	  String ss = files[fileIntList].toString();
+                   	  if (title.equals(ss.substring(ss.lastIndexOf("\\") + 1, ss.length()-4))) {
+                   		  System.out.println("다른 제목을 입력하세요.");
+                   		  title = s4.nextLine();
+                   	  }
+                     }
+                    writer = new BufferedWriter(new FileWriter("c://Example//File//"+ title + ".txt"));
+                    writer.write("날짜:" + date + "\r\n");
+                    writer.write("제목:" + title + "\r\n");
+                     System.out.println("내용을 입력하세요");
+                     String story = s.nextLine();
+                     while (!story.equals("store")){
+                        writer.write( story + "\r\n");
+                        story = s2.nextLine();
+                     }
+                     writer.close();
+                 } catch (FileNotFoundException e) {
+                     e.printStackTrace();
+                 } catch (IOException e) {
+                     e.printStackTrace();
+                 }
+             }
    static int View2_m(){
          File file1 = new File("c://Example//File//");
           File[] files = file1.listFiles();
@@ -124,7 +129,7 @@ public class Main_Memo {
         	  System.out.println("입력한 제목의 메모가 존재하지 않습니다.");
         	  return;
           }
-          
+         
     	 BufferedReader reader = new BufferedReader(new FileReader("c://Example//File//"+ title + ".txt"));
          String data = "";
          
@@ -168,7 +173,12 @@ public class Main_Memo {
          }
          reader.close();
          File file2 = new File("c://Example//File//"+ title+ ".txt");
-         file2.delete();
+         if( file2.exists() ){
+             file2.delete();
+         }
+         if( file2.exists() ){
+             file2.delete();
+         }
          }catch (FileNotFoundException e) {
                e.printStackTrace();
         } catch (IOException e) {
@@ -197,10 +207,21 @@ public class Main_Memo {
     }
    static void Delete_m(String title) {
       File file1 = new File("c://Example//File//"+ title+ ".txt");
-        if( file1.exists() ){
-            file1.delete();
-        }else{
+      if (file1.exists()) {
+              file1.delete();
+      }
+      else {
             System.out.println("파일이 존재하지 않습니다.");
         }
    }
+   
+   static Boolean Deletecheck(String title) {
+	      File file3 = new File("c://Example//File//"+ title+ ".txt");
+	    	  if (file3.exists()) {
+	              file3.delete();
+	    	  }
+	      return true;
+	 }
 }
+
+
